hgapp/requirements.txt
attrs==19.3.0
awsebcli==3.18.1
bcrypt==3.1.7
bleach==2.1.4
blessed==1.9.5
boto3==1.12.49
botocore==1.15.49
cached-property==1.5.1
cement==2.8.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
cryptography==2.9.2
Django==2.2.12
django-appconf==1.0.4
django-bootstrap-form==3.4
django-bootstrap3-datetimepicker-2==2.8.2
django-classy-tags==1.0.0
django-guardian==2.2.0
django-js-asset==1.2.2
django-markdown-deux==1.0.5
django-mptt==0.9.1
django-nyt==1.1.5
django-pagedown==2.1.1
django-postman==3.6.2
django-sekizai==1.1.0
django-storages==1.9.1
django-user-accounts==3.0.2
docker==4.2.0
docker-compose==1.25.5
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.12
future==0.16.0
html5lib==1.0.1
idna==2.7
importlib-metadata==1.6.0
jmespath==0.9.0
jsonfield==3.1.0
jsonschema==3.2.0
Markdown==3.1.1
markdown2==2.3.9
paramiko==2.7.1
pathspec==0.5.9
Pillow==7.1.2
pinax-eventlog==2.0.3
pinax-theme-bootstrap==8.0.1
pinax-webanalytics==3.0.0
psycopg2==2.7.7
pyasn1==0.4.8
pycparser==2.20
PyNaCl==1.3.0
pyrsistent==0.16.0
python-dateutil==2.6.0
pytz==2020.1
PyYAML==5.3.1
requests==2.20.1
rsa==3.4.2
s3transfer==0.3.3
semantic-version==2.5.0
six==1.11.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
termcolor==1.1.0
texttable==0.9.0
urllib3==1.24.3
wcwidth==0.1.9
webencodings==0.5.1
websocket-client==0.37.0
wiki==0.5
zipp==3.1.0
